Store Navigation w/ Arrow + TrapPlacement

+++ NeverQuest/Assets/Scripts/Traps/Trap.cs
@@ -4,8 +4,8 @@ using UnityEngine.UI;
 using UnityEngine;
 
 public class Trap : MonoBehaviour {
-	public int placementX;//so that the position is fine with all of them
-	public int placementY;// some traps would b correctly placed whereas other types wont
+	public float placementX;//so that the position is fine with all of them
+	public float placementY;// some traps would b correctly placed whereas other types wont
 
 	public int cost;
 	public int _damage;
